+++ design-system/stylelint/src/__tests__/utils/_test-rule.tsx
@@ -1,0 +1,8 @@
+/// <reference types="jest-preset-stylelint" />
+
+// @ts-ignore
+import getTestRule from 'jest-preset-stylelint/getTestRule';
+
+const testRule: JestPresetStylelint.TestRule = getTestRule();
+
+export default testRule;