+++ people-and-teams/profilecard/src/components/Team/lazyTeamProfileCard.tsx
@@ -2,6 +2,8 @@ import { lazy } from 'react';
 
 export const TeamProfileCardLazy = lazy(
   () =>
-    // eslint-disable-next-line import/dynamic-import-chunkname
-    import(/* webpackChunkName:"lazy-team-profilecard" */ './TeamProfileCard'),
+    import(
+      /* webpackChunkName: "@atlaskit-internal_lazy-team-profilecard" */
+      './TeamProfileCard'
+    ),
 );