+++ people-and-teams/profilecard/src/components/User/index.ts
@@ -1,9 +1,6 @@
-import { DELAY_MS_HIDE, DELAY_MS_SHOW } from '../../util/config';
+export { DELAY_MS_HIDE, DELAY_MS_SHOW } from '../../util/config';
 
-import ProfileCard from './ProfileCard';
+export { ProfileCardLazy } from './lazyProfileCard';
 import ProfileCardTrigger from './ProfileCardTrigger';
-
-export { DELAY_MS_HIDE, DELAY_MS_SHOW, ProfileCard, ProfileCardTrigger };
-export { default as ProfileCardResourced } from './ProfileCardResourced';
 
 export default ProfileCardTrigger;