+++ people-and-teams/profilecard/src/components/User/lazyProfileCard.tsx
@@ -1,0 +1,9 @@
+import { lazy } from 'react';
+
+export const ProfileCardLazy = lazy(
+  () =>
+    import(
+      /* webpackChunkName: "@atlaskit-internal_lazy-profilecard" */
+      './ProfileCard'
+    ),
+);